fix: use dtype= (canonical) instead of legacy torch_dtype= in from_pretrained
HF transformers docs now recommend `dtype=` over `torch_dtype=` (legacy).
Standardize all AutoModelForCausalLM.from_pretrained calls across the codebase.

+++ libs/model-training/src/model_training/adapter_generator.py
@@ -89,7 +89,7 @@ def extract_activations(
 
     model: Any = AutoModelForCausalLM.from_pretrained(
         base_model_name,
-        torch_dtype=activation_dtype,
+        dtype=activation_dtype,
     )
     model = model.to(device)  # type: ignore[assignment]
     model.eval()
